+++ log4stdc/logobjects.h
@@ -189,7 +189,7 @@ int  l4sc_default_clone_object(l4sc_objcptr_t obj, void *buf, size_t bufsize);
 size_t l4sc_default_get_object_size(l4sc_objcptr_t obj);
 unsigned l4sc_default_get_object_hashcode(l4sc_objcptr_t obj);
 int l4sc_default_is_equal_object(l4sc_objcptr_t obj, l4sc_objcptr_t other);
-int l4sc_default_get_object_length(l4sc_objcptr_t obj);
+size_t l4sc_default_get_object_length(l4sc_objcptr_t obj);
 int l4sc_default_object_tostring(l4sc_objcptr_t obj, char *buf, size_t bufsize);
 
 void l4sc_default_set_object_name(l4sc_objptr_t obj, const char *name, int len);
