+++ mempools/stdcpool.c
@@ -16,7 +16,7 @@ static int  clone_pool(const struct mempool *pool, void *buf, size_t bufsize);
 static size_t get_pool_object_size(const struct mempool *pool);
 static unsigned get_pool_hashcode(const struct mempool *pool);
 static int is_equal_pool(const struct mempool *obj,const struct mempool *other);
-static int get_pool_size(const struct mempool *pool);
+static size_t get_pool_size(const struct mempool *pool);
 static int pool_tostring(const struct mempool *pool, char *buf, size_t bufsize);
 
 /*
@@ -122,7 +122,7 @@ is_equal_pool(const struct mempool *obj, const struct mempool *other)
 	return (obj == other);
 }
 
-static int
+static size_t
 get_pool_size(const struct mempool *pool)
 {
 	return (0);
